requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.23
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
chromadb>=0.4.18
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.1.0
diffusers>=0.24.0
accelerate>=0.25.0
Pillow>=10.1.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.3.0
spotipy>=2.22.1
numpy>=1.24.3
pandas>=2.0.3
